checklist+update control keydown function int to WORD

+++ Common/Control.h
@@ -43,7 +43,7 @@ public:
 	// virual function implemnted by every component
 	virtual void draw(Graphics& g, short x, short y, size_t z) ;
 	virtual void mousePressed(int x, int y, bool isLeft) {};
-	virtual void keyDown(int keyCode, char charecter) {};
+	virtual void keyDown(WORD keyCode, char charecter) {};
 
 	virtual void getAllControls(vector<Control*>* controls) {};
 	virtual bool canGetFocus() { return FALSE; };
